Add stats collection for spray client
