+++ README.md
@@ -24,11 +24,11 @@ class Counter(val counterName: String) extends Actor {
   var hours = 0L
 
   lazy val stats = context.actorOf(Stats.props())
-  val secondsCounter = Increment("seconds")
-  val minutesCounter = Increment("minutes")
-  val hoursCounter = Increment("hours")
-  val gaugeMetric = Gauge("shotgun")(12L)
-  val timingMetric = Timing("tempo")(5.seconds)
+  val secondsCounter = Increment(Bucket("seconds"))
+  val minutesCounter = Increment(Bucket("minutes"))
+  val hoursCounter = Increment(Bucket("hours"))
+  val gaugeMetric = Gauge(Bucket("shotgun"))(12L)
+  val timingMetric = Timing(Bucket("tempo"))(5.seconds)
 
   def receive = {
     case IncrementSecond =>
@@ -51,7 +51,7 @@ class Counter(val counterName: String) extends Actor {
 The counter messages themselves are curried for convenience:
 
 ```scala
-val interval = Timing("interval")
+val interval = Timing(Bucket("interval"))
 stats ! interval(3.seconds)
 stats ! interval(2.seconds)
 stats ! interval(1.second)
@@ -62,7 +62,7 @@ If a `Stats` instance is assigned a namespace(via the Config object) then all co
 ```scala
 val page1Perf = system.actorOf(Stats.props(<config_with_namespace>))
 val page2Perf = system.actorOf(Stats.props(<config_with_namespace>))
-val response = Timing("response")
+val response = Timing(Bucket("response"))
 page1Perf ! response(250.milliseconds)
 page2Perf ! response(100.milliseconds)
 ```
@@ -71,8 +71,8 @@ But you don't have to use namespaced actors at all:
 
 ```scala
 val perf = system.actorOf(Stats.props())
-perf ! Timing("page1.response")(250.milliseconds)
-perf ! Timing("page2.response")(100.milliseconds)
+perf ! Timing(Bucket("page1.response"))(250.milliseconds)
+perf ! Timing(Bucket("page2.response"))(100.milliseconds)
 ```
 
 ### API
@@ -82,10 +82,10 @@ There is a simple api if you dont want to use the actor's `!` method's
 ```scala
 implicit val statsActor = context.actorOf(Stats.props())
 
-Stats.increment("endpoint1")
-Stats.timing("page2.response")(250 milliseconds)
+Stats.increment(Bucket("endpoint1"))
+Stats.timing(Bucket("page2.response"))(250 milliseconds)
 
-Stats.withTimer("code.execution.time") {
+Stats.withTimer(Bucket("code.execution.time")) {
     //execute code here
     Thread.sleep(new Random().nextInt())
 }
@@ -100,13 +100,19 @@ Releases and snapshots are hosted on The New Motion public repository. To add de
 ```scala
 resolvers += "The New Motion Public Repo" at "http://nexus.thenewmotion.com/content/groups/public/"
 
-libraryDependencies += "com.thenewmotion" %% "akka-statsd-core" % "1.0.0"
+libraryDependencies += "com.thenewmotion" %% "akka-statsd-core" % "0.9.0"
+```
+
+For stats collection over HTTP requests served by spray server or issued by spray client, use respective dependencies from below:
+```
+libraryDependencies += "com.thenewmotion" %% "akka-statsd-spray-server" % "0.9.0"
+libraryDependencies += "com.thenewmotion" %% "akka-statsd-spray-client" % "0.9.0"
 ```
 
 
 ## Explanation
 
-This implementation is intended to high performance, thus it
+This implementation is intended to be of high performance, thus it
 
 - uses an akka.io UdpConnected implementation to cache local security checks within the JVM if SecurityManager is enabled
 - allows for message batching before sending them out to StatsD, as per [statsd Multi-Metric Packets](https://github.com/etsy/statsd/blob/master/docs/metric_types.md#multi-metric-packets) specification. Turned on by default
@@ -127,11 +133,33 @@ As messages are transmitted to a stats actor those messages are queued for later
 
 This can be turned off in the configuration by setting `enable-multi-metric = false`
 
+## Bucket transformations
+
+Any `UUID` in the bucket path substituted with token `[id]`. In addition to this, clients can specify custom transformations to influence the way bucket names are augmented before being sent to StatsD server.
+
+In order to do so, the client configuration should contain `akka.statsd.transformations` section of the following format:
+
+"""
+akka.statsd.transformations = [
+  {
+    pattern = "/foo/[a-z0-9]+/bar",
+    into    = "/foo/[segment]/bar"
+  }
+]
+"""
+
+This reads as: if the part of the bucket matches the regular expression in `pattern`, replace it with the one described in `into` (which is *not* a regular expression).
+
+Please note that:
+- client-defined transformations take priority over `UUID` replacement
+- transformations are matched from top to bottom
+- every transformation that matches the path will be applied
+
 
 ## Configuration
 
 By default `Stats.props()` uses `Config` which is constructed by resolving Typesafe Config with call to `ConfigFactory.load()`.
-You may use `deploymentzone.Config`, constructed manually or from provided Typesafe Config instance too. 
+You may use `akkas.statsd.Config`, constructed manually or from provided Typesafe Config instance too. 
 Here are the default settings, with the exception of hostname, which is a required setting:
 
 ```
@@ -145,9 +173,22 @@ akka.statsd {
         # commodity internet:    512
         packet-size = 1432
         transmit-interval = 100 ms
+
+        transformations = [
+          {
+            pattern = "foo"
+            into = "bar"
+          }
+        ]
     }
 }
 ```
+
+## Compatibility
+
+Since version `0.9.0`, if typesafe configuration used, provide settings in namespace `akka.statsd`.
+
+Versions before `0.9.0` use `deploymentzone.statsd`
 
 ## Influences
