+++ akka-statsd-core/src/main/resources/reference.conf
@@ -18,5 +18,11 @@ akka.statsd {
 
   enable-multi-metric = true
 
-  transform = []
+  # must be of form
+  # {
+  #   pattern = "some regular expression"
+  #   into = "some string"
+  # }
+  transformations = []
+
 }
