--- akka-statsd-core/src/main/scala/deploymentzone/Transform.scala
@@ -1,6 +1,0 @@
-package deploymentzone
-
-import scala.util.matching.Regex
-
-
-case class Transform(pattern: Regex, into: String)
