--- akka-statsd-core/src/main/scala/deploymentzone/actor/domain/NamespaceTransformer.scala
@@ -1,28 +1,0 @@
-package deploymentzone.actor.domain
-
-import deploymentzone.actor.Metric
-import deploymentzone.actor.validation.StatsDBucketValidator
-
-/**
- * Transforms the toString result value of a CounterMessage instance to include an
- * optional namespace.
- */
-private[actor] class NamespaceTransformer(val namespace: String) extends ((Metric[_]) => String) {
-
-  require(StatsDBucketValidator(namespace))
-
-  override def apply(counter: Metric[_]): String = {
-    require(counter != null)
-
-    val intermediate = counter.toString
-    namespace match {
-      case null => intermediate
-      case "" => intermediate
-      case _ => s"$namespace.$intermediate"
-    }
-  }
-}
-
-private[actor] object NamespaceTransformer {
-  def apply(namespace: String) = new NamespaceTransformer(namespace)
-}
